Restructured files. No real code changes.
