+++ opensourceleg/hardware/thermal.py
