+++ tests/test_actuators/test_gains.py
@@ -1,16 +1,16 @@
 import pytest
 
-from opensourceleg import control
+from opensourceleg.hardware.actuators import Gains
 
 # Defining the Gains objects
-zero_gains = control.Gains()
-nonzero_gains = control.Gains()
+zero_gains = Gains()
+nonzero_gains = Gains()
 nonzero_gains.kp = 1
 nonzero_gains.ki = 2
 nonzero_gains.kd = 3
 nonzero_gains.K = 4
 nonzero_gains.B = 5
-negative_gains = control.Gains()
+negative_gains = Gains()
 negative_gains.kp = -1
 negative_gains.ki = -2
 negative_gains.kd = -3
