+++ tests/test_utilities/test_loops.py
@@ -2,7 +2,7 @@ import time
 
 import pytest
 
-from opensourceleg.utilities import LoopKiller, SoftRealtimeLoop
+from opensourceleg.tools.utilities import LoopKiller, SoftRealtimeLoop
 from tests.test_joints.test_joint import patch_time_time
