+++ tutorials/impedance_mode.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-import opensourceleg.units as units
+import opensourceleg.tools.units as units
 from opensourceleg.osl import OpenSourceLeg
 
 osl = OpenSourceLeg(frequency=200)  # 200 Hz
